create userController
